backend/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
certifi==2019.11.28
cffi==1.14.0
cryptography==2.9
distlib==0.3.0
Django==3.0.5
django-cors-headers==3.2.1
django-nested-admin==3.2.4
django-rest-knox==4.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
filelock==3.0.12
pipenv==2018.11.26
psycopg2==2.8.4
pycparser==2.20
PyJWT==1.7.1
python-monkey-business==1.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
virtualenv==20.0.10
virtualenv-clone==0.5.3
whitenoise==5.0.1
